[Refactor]: 메인 화면 코드 구조 리팩토링
- Models: MainModels.swift로 모델 분리 (RelayList, Song, ChartSong 등)
- Client: MainClient.swift로 Dependency 분리
- Client: MainMockData.swift로 목데이터 분리
- Components: 섹션별 뷰 컴포넌트 분리
  - MainAppBar, RelayListSection, ChartSection
  - BannerSection, ArtistRankingSection
  - PlaylistSection, YouTubeSection
- Extensions: Color+Hex.swift로 공용 Color extension 분리
- MainFeature.swift: Reducer 로직만 유지
- MainView.swift: 메인 뷰 조합만 담당
